backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
supabase>=2.10.0
google-generativeai>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
numpy>=1.26.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
websockets>=13.0,<16
